Change input position to single percision

+++ graphic_helpers/shaders/vs.vert.glsl
@@ -1,6 +1,6 @@
 #version 460 core
 
-layout(location = 0) in dvec3 in_pos;
+layout(location = 0) in vec3 in_pos;
 layout(location = 1) in vec3 in_color;
 layout(location = 2) uniform mat4x4 in_MVP;
 out vec3 vertex_color;
